+++ crates/hyperdrive-math/src/lp.rs
@@ -1,4 +1,4 @@
 mod add;
-mod helpers;
 mod math;
 mod remove;
+mod utils;
