+++ crates/hyperdrive-math/src/lp/utils.rs
